Move DnssecSummary to proto crate
